requirements.txt
Flask==2.3.2
customtkinter==5.1.1
pandas==2.0.3
python-pptx==0.6.21
nest_asyncio==1.5.7
scrapegraphai==1.2.0  # Replace with the correct version if known
google-crc32c==1.5.0